requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
bitsandbytes>=0.40.0
xformers>=0.0.20
regex 
tqdm
scipy
pandas